connector.php
If we want to design our own HTML button or styles, then we have to edit the existing button. This can be a tough style HTML change since it's hardcoded and generated on a page load. The change basically should make it easier to style the button, by checking if the ID "sso_button" already exists on the login, if so, then don't generate the button, so we don't get two buttons on the Login screen.